+++ threedi_modelchecker/checks/other.py
@@ -235,8 +235,8 @@ class Use0DFlowCheck(BaseCheck):
 
     def description(self):
         return (
-            "When %s is used, there should exist at least one "
-            "(impervious) surface." % self.column
+            f"When {self.column_name} is used, there should exist at least one "
+            "(impervious) surface."
         )
 
 
@@ -705,6 +705,104 @@ class PumpStorageTimestepCheck(BaseCheck):
 
     def description(self) -> str:
         return f"{self.column_name} will empty its storage faster than one timestep, which can cause simulation instabilities"
+
+
+class ImperviousNodeInflowAreaCheck(BaseCheck):
+    """Check that total inflow area per connection node is no larger than 10000 square metres"""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(column=models.ConnectionNode.id, *args, **kwargs)
+
+    def get_invalid(self, session: Session) -> List[NamedTuple]:
+        impervious_surfaces = (
+            select(models.ImperviousSurfaceMap.connection_node_id)
+            .select_from(models.ImperviousSurfaceMap)
+            .join(
+                models.ImperviousSurface,
+                models.ImperviousSurfaceMap.impervious_surface_id
+                == models.ImperviousSurface.id,
+            )
+            .group_by(models.ImperviousSurfaceMap.connection_node_id)
+            .having(func.sum(models.ImperviousSurface.area) > 10000)
+        ).subquery()
+
+        return (
+            session.query(models.ConnectionNode)
+            .filter(
+                models.ConnectionNode.id == impervious_surfaces.c.connection_node_id
+            )
+            .all()
+        )
+
+    def description(self) -> str:
+        return f"{self.column_name} has a an associated inflow area larger than 10000 m2; this might be an error."
+
+
+class PerviousNodeInflowAreaCheck(BaseCheck):
+    """Check that total inflow area per connection node is no larger than 10000 square metres"""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(column=models.ConnectionNode.id, *args, **kwargs)
+
+    def get_invalid(self, session: Session) -> List[NamedTuple]:
+        pervious_surfaces = (
+            select(models.SurfaceMap.connection_node_id)
+            .select_from(models.SurfaceMap)
+            .join(
+                models.Surface,
+                models.SurfaceMap.surface_id == models.Surface.id,
+            )
+            .group_by(models.SurfaceMap.connection_node_id)
+            .having(func.sum(models.Surface.area) > 10000)
+        ).subquery()
+
+        return (
+            session.query(models.ConnectionNode)
+            .filter(models.ConnectionNode.id == pervious_surfaces.c.connection_node_id)
+            .all()
+        )
+
+    def description(self) -> str:
+        return f"{self.column_name} has a an associated inflow area larger than 10000 m2; this might be an error."
+
+
+class NodeSurfaceConnectionsCheck(BaseCheck):
+    """Check that no more than 50 surfaces are mapped to a connection node"""
+
+    def __init__(
+        self,
+        check_type: Literal["impervious", "pervious"] = "impervious",
+        *args,
+        **kwargs,
+    ):
+        super().__init__(column=models.ConnectionNode.id, *args, **kwargs)
+
+        self.surface_column = None
+        if check_type == "impervious":
+            self.surface_column = models.ImperviousSurfaceMap
+        elif check_type == "pervious":
+            self.surface_column = models.SurfaceMap
+
+    def get_invalid(self, session: Session) -> List[NamedTuple]:
+        if self.surface_column is None:
+            return []
+
+        overloaded_connections = (
+            select(self.surface_column.connection_node_id)
+            .group_by(self.surface_column.connection_node_id)
+            .having(func.count(self.surface_column.connection_node_id) > 50)
+        ).subquery()
+
+        return (
+            session.query(models.ConnectionNode)
+            .filter(
+                models.ConnectionNode.id == overloaded_connections.c.connection_node_id
+            )
+            .all()
+        )
+
+    def description(self) -> str:
+        return f"{self.column_name} has more than 50 surface areas mapped to it; this might be an error."
 
 
 class BetaColumnsCheck(BaseCheck):